requirements.txt
PyQt6==6.6.1
pyinstaller==6.2.0
pyserial==3.5
requests==2.31.0
python-dateutil==2.8.2
obd==0.7.1
pybluez==0.23
